Gmail Update

+++ src/main/java/com/bridgeit/todoApplication/UserDAO/UserDAO.java
@@ -11,18 +11,18 @@ public interface UserDAO {
 		
 		void addEntity(User user) throws Exception;
 
-		User getEntityById(int id) throws Exception;
+		User getEntityById(long id) throws Exception;
 
 		List<User> getUserList() throws Exception;
 
-		void deleteEntity(int id) throws Exception;
+		void deleteEntity(long id) throws Exception;
 
 		User authUser(String email, String password);
 
 		User getEntityByEmailId(String email);
 
 		void savePicture(UserPicture picture);
-		UserPicture getPicture(int userId);
+		UserPicture getPicture(long userId);
 
 	}
